# laravel-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/categories/{id}/products": {
            "get": {
                "tags": [
                    "Authorize"
                ],
                "summary": "Get products by category",
                "description": "Return a list of products that belong to category",
                "operationId": "b938ef418ed36bf03c6189a77c957f85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResourceCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Authorize"
                ],
                "summary": "Get product by id",
                "description": "Returns a product with same id",
                "operationId": "b46f4aa73232d0368eee698a082342a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 87
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "22"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "3"
                    },
                    "name": {
                        "type": "string",
                        "example": "coll laptop gt6"
                    },
                    "description": {
                        "type": "string",
                        "example": "some text goes here"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": "1488.92"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://logos-world.net/wp-content/uploads/2020/11/Fox-News-Channel-Emblem.png"
                    }
                },
                "type": "object"
            },
            "ProductResourceCollection": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}